Move TileMap to _base.py
